fix: cast mask to int dtype in load_tiff_timeseries to prevent float-label crashes

+++ trackastra/model/model_api.py
@@ -382,7 +382,7 @@ class Trackastra:
         if masks_path.is_dir():
             masks = load_tiff_timeseries(masks_path)
         else:
-            masks = tifffile.imread(masks_path)
+            masks = tifffile.imread(masks_path).astype(np.int32)
 
         if len(imgs) != len(masks):
             raise RuntimeError(
